minor refactoring
move negotiation backend methods to a mixin to
unclutter the client class. Also group the negotiation specs
to keep them together when we add more client functionality

MS-2527
